requirements.txt
streamlit>=1.28.0
azure-search-documents>=11.4.0
azure-ai-formrecognizer>=3.3.0
azure-cosmos>=4.5.0
azure-identity>=1.15.0
openai>=1.6.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
requests>=2.31.0
azure-ai-evaluation>=1.0.0
azure-ai-projects>=2.0.0b1
azure-monitor-opentelemetry>=1.2.0
azure-core-tracing-opentelemetry>=1.0.0
opentelemetry-api>=1.37.0
opentelemetry-sdk>=1.37.0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-instrumentation-openai-v2>=1.0.0
python-multipart>=0.0.6